feat: findIndex
